+++ R/is_local.R
@@ -2,12 +2,17 @@
 #' 
 #' Determine whether a file exist locally.
 #' @param path Path to local file or remote URL.
+#' @inheritParams is_url
 #' @export
 #' @examples
 #' echodata::is_local(
 #'     path="https://github.com/RajLabMSSM/echotabix/raw/main/README.md"
 #' )
-is_local <- function(path) {
+is_local <- function(path,
+                     protocols=c("http","https",
+                                 "ftp","ftps",
+                                 "fttp","fttps")) {
     # ssh.utils::file.exists.remote(file = path) # Doesn't work?
-    (!is_url(path)) && (file.exists(path))
+    (!is_url(fileName = path, 
+             protocols = protocols)) && (file.exists(path))
 }