+++ tests/testthat/test-is_url.R
@@ -1,0 +1,9 @@
+test_that("is_url works", {
+  
+    
+    fileName <- "https://github.com/RajLabMSSM/echolocatoR/raw/master/NEWS.md"
+    testthat::expect_true(is_url(fileName=fileName))
+
+    fileName <- "://github.com/RajLabMSSM/echolocatoR/raw/master/NEWS.md"    
+    testthat::expect_false(is_url(fileName=fileName))
+})